Release v2.6.0: GATA Project & DTO Editing Infrastructure
Major Features:
- Event-driven cache invalidation system with EventBus integration
- Model-independent DTO-based parameter editing (tools/update_model_parameters.py)
- GATA1/PU.1 bistable switch model validation (hematopoietic lineage commitment)
- Programmatic model editing with automatic cache invalidation

Documentation:
- GATA project complete documentation (workspace/projects/gata/README.md)
- Programmatic model editing guide (doc/PROGRAMMATIC_MODEL_EDITING.md)
- Model independence architecture (doc/MODEL_INDEPENDENCE.md)
- Event-driven cache system (doc/EVENT_DRIVEN_CACHE_INVALIDATION.md)

Repository Organization:
- Scripts moved to scripts/ directory (analyze_gata, visualize, factorial analysis)
- Tests organized with new validation tests (test_event_cache_invalidation.py)
- Tools directory with model editing utilities
- All README files updated and cross-referenced

GATA Model Validation:
- Energy metabolism fixes (ATP synthesis 10→10,000, GTP regeneration 500)
- Transcription rate fixes (1.2→0.08, 15× slower)
- Degradation rate fixes (0.01→0.05/0.015→0.075, 5× faster)
- Realistic biological parameters validated (EPO=150, 1000s)

Architecture:
- ModelRepository refactored to event-driven pattern
- DocumentModel emits 'file.saved' events
- Cache invalidation on external edits
- DTO property setters with validation

API Version: 2.6

+++ src/shypn/ui/panels/viability/ui/subnet_parameters_view.py
@@ -108,8 +108,8 @@ class SubnetParametersView(Gtk.Box):
         
         Columns: ID, Name, Marking (editable), Type, Label, Background, BackgroundSet
         """
-        # Create ListStore: id, name, marking (int, editable), type, label, background, background_set
-        store = Gtk.ListStore(str, str, int, str, str, str, bool)
+        # Create ListStore: id, name, marking (float, editable), type, label, background, background_set
+        store = Gtk.ListStore(str, str, float, str, str, str, bool)
         
         # Create TreeView
         treeview = Gtk.TreeView(model=store)
